refactor: deduplicate SDK testing files between v0.6 and v0.7 (QS-040)
mock-context.ts, mock-agents.ts, and test-harness.ts were 100% identical
between v0.6 and v0.7 plugin-testing packages. Bug fixes required updating
both copies. Extracted to sdk/shared/plugin-testing/ and replaced both
version copies with single-line re-exports. mock-api.ts remains separate
as it has version-specific differences (FilesAPI.watch, HubAPI.refresh,
ALL_SOUND_EVENTS).
