+++ sdk/shared/plugin-testing/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@clubhouse/plugin-testing-shared",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Shared testing utilities for Clubhouse plugin SDK versions",
+  "exports": {
+    "./mock-context": "./src/mock-context.ts",
+    "./mock-agents": "./src/mock-agents.ts",
+    "./test-harness": "./src/test-harness.ts"
+  },
+  "peerDependencies": {
+    "@clubhouse/plugin-types": ">=0.6.0"
+  }
+}
