+++ sdk/shared/plugin-testing/src/mock-context.ts
@@ -1,0 +1,33 @@
+import { tmpdir } from "node:os";
+import { join } from "node:path";
+import type { PluginContext, PluginScope } from "@clubhouse/plugin-types";
+
+interface MockContextOptions {
+  pluginId?: string;
+  pluginPath?: string;
+  projectId?: string;
+  projectPath?: string;
+  scope?: PluginScope;
+  settings?: Record<string, unknown>;
+}
+
+/**
+ * Creates a `PluginContext` with sensible defaults for testing.
+ *
+ * ```ts
+ * const ctx = createMockContext({ pluginId: "my-plugin" });
+ * activate(ctx, api);
+ * expect(ctx.subscriptions.length).toBeGreaterThan(0);
+ * ```
+ */
+export function createMockContext(options?: MockContextOptions): PluginContext {
+  return {
+    pluginId: options?.pluginId ?? "test-plugin",
+    pluginPath: options?.pluginPath ?? join(tmpdir(), "plugins", "test-plugin"),
+    projectId: options?.projectId ?? "test-project",
+    projectPath: options?.projectPath ?? join(tmpdir(), "test-project"),
+    scope: options?.scope ?? "project",
+    subscriptions: [],
+    settings: options?.settings ?? {},
+  };
+}
