+++ sdk/v0.6/plugin-testing/package.json
@@ -21,6 +21,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@clubhouse/plugin-testing-shared": "file:../../shared/plugin-testing",
     "@clubhouse/plugin-types": "file:../plugin-types"
   },
   "peerDependencies": {
