+++ sdk/v0.7/plugin-testing/package.json
@@ -22,6 +22,7 @@
     "test": "vitest run"
   },
   "dependencies": {
+    "@clubhouse/plugin-testing-shared": "file:../../shared/plugin-testing",
     "@clubhouse/plugin-types": "file:../plugin-types"
   },
   "peerDependencies": {
